requirements.txt
pika
logging
gensim
NLTK
stop_words
neo4j-driver